requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
fsspec==0.8.7
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
googledrivedownloader==0.4
grpcio==1.36.1
idna==2.10
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
networkx==2.5
nibabel==3.2.1
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pathtools==0.1.2
Pillow==8.1.2
promise==2.3
protobuf==3.15.6
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rising==0.2.0.post0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
sentry-sdk==1.0.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.3.5
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.10.22
Werkzeug==1.0.1
yarl==1.6.3
